Removed SetGlyphCache from GraphicContext Removed GetGlyphCache altogether Added NewGlyphCache to draw2dbase
